Refactored the upload and moved the upload functionality to another file
