+++ thinkback/models.py
@@ -22,7 +22,7 @@ class Problem:
 		self.desc = desc
 		self.function = function
 		
-class UploadedFile:
+class ProblemModule:
 	def __init__(self, file):
 		self.file = file
 		self._allowed_extensions = set(['py'])
@@ -32,6 +32,7 @@ class UploadedFile:
 			return True
 		return False
 
+
 	def is_allowed(self):
 		return '.' in self.file.filename and \
 			self.file.filename.rsplit('.', 1)[1].lower() in self._allowed_extensions
@@ -39,7 +40,7 @@ class UploadedFile:
 	def create_file_path(self, problem_id):
 		return os.path.join(app.config['UPLOAD_FOLDER'], ''.join(problem_id))
 
-	def save_file_to_path(self, path, secure_filename):
+	def save_files_to_path(self, path, secure_filename):
 		self.file.save(os.path.join(path, secure_filename))
 		self.file.save(os.path.join(path, '__init__.py'))
