+++ thinkback/templates/problem.html
@@ -10,7 +10,7 @@
 		<h2>
 			{{problem.name, problem.desc}}
 		</h2>
-		<form action="{{ url_for('/assignments.upload_file', link=link, problem_id=problem['id'] ) }}" method="POST" enctype="multipart/form-data">
+		<form action="{{ url_for('upload.upload_file', link=link, problem_id=problem['id'] ) }}" method="POST" enctype="multipart/form-data">
 			<input type="file" name="file">
 			<input type="submit" value="Submit file">
 		</form>
