+++ thinkback/views/assignments.py
@@ -1,20 +1,16 @@
 # thinkback/views/assignments.py
 import os
+import json
 import sqlite3
 import importlib
-from ..models import Assignment, Problem, UploadedFile
 from flask import current_app as app
 from werkzeug.utils import secure_filename
+from ..models import Assignment, Problem, ProblemModule
 from flask import Blueprint, render_template, request, redirect, flash, url_for, g
-from pprint import pprint
-import json
 
 ALLOWED_EXTENSIONS = set(['py'])
 
 assignment_blueprint = Blueprint('/assignments', __name__)
-
-# Gives the right path to the 'link' coming in, and also gets a right question list
-
 
 @assignment_blueprint.route('/<link>', methods=['GET'])
 def get_assignment_path(link):
@@ -37,42 +33,6 @@ def get_problems(link, problem_id):
 	problem = get_single_problem(problem_id)
 	return render_template('problem.html', link=link, problem=problem, results={})
 
-
-@assignment_blueprint.route('/<link>/<problem_id>', methods=['POST'])
-def upload_file(link, problem_id):
-	if request.method == 'POST':
-		# Check if the post request has a file in it
-		file = request.files['file']
-		submitted_file = UploadedFile(file)
-		if submitted_file.is_empty():
-			flash("Something went wrong")
-			return redirect(request.url)
-		if submitted_file.file and submitted_file.is_allowed():
-			filename = secure_filename(file.filename)
-			path = submitted_file.create_file_path(problem_id)
-			if not os.path.exists(path):
-				os.makedirs(path)
-			submitted_file.save_file_to_path(path, filename)
-			module_path = '.uploads.{}'.format(problem_id)
-			solution_path = '.impl.{}'.format(problem_id)
-			problem_module = submitted_file.get_file_module(
-				module_path, filename)
-			function_name = get_function_name(problem_id)
-			# Try to import the fuction needed for the problem
-			try:
-				problem_function = getattr(problem_module, function_name)
-				solution = importlib.import_module('.correct', package=solution_path) 
-				tmp = solution.Solution(problem_function)
-				results = tmp.run_tests()
-				problem = get_single_problem(problem_id)
-				return render_template('problem.html', link=link, problem=problem, results=results)
-			except AttributeError:
-				# If the function did not exists remove the file
-				os.remove(os.path.join(path, filename))
-			return redirect(request.url)
-
-	# TODO: Return error that something went wrong
-	return ""
 
 
 def filter_assignments(assignments, status):
@@ -113,15 +73,6 @@ def get_single_problem(problem_id):
 	problem = Problem(entry['p_id'], entry['a_id'], entry['p_name'],
 					  entry['p_desc'], entry['p_solution_name'])
 	return problem
-
-
-def get_function_name(problem_id):
-	db = get_db()
-	cur = db.execute(
-		'select P.p_solution_name from Problems P where P.p_id = {}'.format(problem_id))
-	entry = cur.fetchone()
-	return entry['p_solution_name']
-
 
 def connect_db():
 	print("""Connects to the specific database.""")
